Change Enter Information Page and Create Preview Information Page
The inputs entered in the Enter Information page will be reflected on the Preview Page. In the Preview Page will have the Submit Form and Edit Information button. If the user chooses to submit form, a confirmation message will alert. If the user chooses to edit information, the previous information he/she entered will be preserved.